pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]  
build-backend = "setuptools.build_meta"


[project]
name = "yoso"
version = "0.0.1"
description = "You only shoot once."
authors = [
  { name = "Trinh Quoc Anh", email = "trinhquocanh94@gmail.com" },
]
license = { text = "CC-BY-NC-SA-4.0" }
readme = "README.md"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = "==3.10.*"

dependencies = [
    # PyTorch stack (CUDA 11.8)
    "torch==2.0.1+cu118",
    "torchvision==0.15.2+cu118",
    "torchaudio==2.0.2+cu118",

    # HF ecosystem (compatible versions)
    "transformers==4.36.2",
    "diffusers==0.25.0",
    "huggingface-hub>=0.19.3,<0.26",

    # Training & metrics
    "accelerate==0.25.0",
    "torchmetrics==1.2.1",

    # Utils & infra
    "tqdm==4.66.1",
    "einops==0.7.0",
    "bitsandbytes==0.39.0",
    "scipy==1.11.1",
    "opencv-python",
    "gradio==4.24.0",
    "fvcore",
    "cloudpickle",
    "omegaconf",
    "pycocotools",
    "basicsr",
    "av",
    "onnxruntime==1.23.0",
    "pydantic==2.10.6",
    "scikit-image",
    "colour-science",
    "lpips"
]

[tool.setuptools]
packages = ["yoso", "densepose", "detectron2"]

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.repo-review]
ignore = ["PY004", "PY007", "PP003", "PP301", "GH102", "GH103", "GH212", "PC", "MY", "RF", "RTD"]

[tool.ruff]
select = ["E", "F"]
ignore = ["E501", "E731","E402", "E721", "F401", "E741", "F841", "F822", "E711", "F821", "F405","F403","E712"]
line-length = 88